chore(all): clean up react typescript structure and props
