+++ packages/grid-react/src/GridElement.tsx
@@ -1,8 +1,19 @@
-import React from "react";
-import { GridElementProps } from "./internal/types";
-import { getColumnsByScreenSize } from "./internal/getColumnsByScreenSize";
+import React, { ReactNode } from "react";
+import { getColumnsByScreenSize } from "./getColumnsByScreenSize";
 
-export const GridElement = ({ children, columnSpan, className = "" }: GridElementProps) => {
+export interface ColumnType {
+    small?: number;
+    medium?: number;
+    large?: number;
+}
+
+interface Props {
+    children?: ReactNode;
+    className?: string;
+    columnSpan: number | ColumnType;
+}
+
+export const GridElement = ({ children, columnSpan, className = "" }: Props) => {
     const columnClass =
         typeof columnSpan === "number" ? `jkl-grid__element--${columnSpan}` : getColumnsByScreenSize(columnSpan);
     return <div className={`${columnClass} ${className}`}>{children}</div>;
