--- packages/logo-react/src/internal/types.ts
@@ -1,5 +1,0 @@
-export interface LogoProps {
-    width?: number;
-    positive?: boolean;
-    className?: string;
-}
